Implementing permission system for users. Adding utils file for different utilities. Correcting some bugs.

+++ Epic_Events/CustomUser/management/commands/user.py
@@ -3,11 +3,13 @@ from django.contrib.auth.models import Permission
 from django.db import transaction
 from getpass import getpass
 from CustomUser.models import CustomUserAccount, Team
+from CustomUser.permissions import IsAuthenticated, IsSuperuser, IsSameUser
 
 
-USER_FIELDS = ['username', 'email', 'first_name', 'last_name', 'phone_number', 'team_name']
+USER_FIELDS = ['current_user', 'username', 'email', 'first_name', 'last_name', 'phone_number', 'team_name']
 
 USER_DESCRIPTIONS = {
+    'current_user': "Enter current user's username: ",
     'username': "Enter username: ",
     'email': "Enter email: ",
     'first_name': "Enter first name: ",
@@ -16,6 +18,7 @@ USER_DESCRIPTIONS = {
     'team_name': "Enter team name: ",
 }
 
+valid_team_names = ['sales', 'management', 'support']
 
 class Command(BaseCommand):
     help = 'Command lines to manage CRUD operations on Users from the CustomUserAccount model.'
@@ -47,25 +50,39 @@ class Command(BaseCommand):
         else:
             raise CommandError('Invalid command')
 
-def list_users(self):
-    try:
-        users = CustomUserAccount.objects.all()
-        if not users:
-            self.stdout.write('No users exist.')
-        else:
-            for user in users:
-                self.stdout.write(str(user))
-    except Exception as e:
-        self.stdout.write('An error occurred: {}'.format(e))
+    def list_users(self):
+
+        try:
+            users = CustomUserAccount.objects.all()
+            if not users:
+                self.stdout.write('No users exist.')
+            else:
+                for user in users:
+                    self.stdout.write(f'Username: {user.username}, First Name: {user.first_name}, Last Name: {user.last_name}, Last Login: {user.last_login}')
+        except Exception as e:
+            self.stdout.write('An error occurred: {}'.format(e))
 
     def create_user(self, options):
-        username = options['username'] or input('Enter username: ')
-        email = options['email'] or input('Enter email: ')
-        first_name = options['first_name'] or input('Enter first name: ')
-        last_name = options['last_name'] or input('Enter last name: ')
-        phone_number = options['phone_number'] or input('Enter phone number: ')
+        current_user_name = options['current_user'] or input(USER_DESCRIPTIONS['current_user'])
+        username = options['username'] or input(USER_DESCRIPTIONS['username'])
+        email = options['email'] or input(USER_DESCRIPTIONS['email'])
+        first_name = options['first_name'] or input(USER_DESCRIPTIONS['first_name'])
+        last_name = options['last_name'] or input(USER_DESCRIPTIONS['last_name'])
+        phone_number = options['phone_number'] or input(USER_DESCRIPTIONS['phone_number'])
         password = getpass('Enter password: ')
-        team_name = options['team_name'] or input('Enter team name: ')
+        team_name = (options['team_name'] or input(USER_DESCRIPTIONS['team_name'])).lower()
+        while team_name not in valid_team_names:
+            print(f'Invalid team name: {team_name}. Valid team_name options are : {", ".join(valid_team_names)}')
+            team_name = input(USER_DESCRIPTIONS['team_name']).lower()
+        
+        try:
+            current_user = CustomUserAccount.objects.get(username=current_user_name)
+            permission = IsAuthenticated(current_user).has_permission() and (IsSuperuser(current_user).has_permission() or IsManager(current_user).has_permission)
+            if not permission:
+                raise CommandError('Current user is not authenticated and/or does not have permission to create a new user')
+        except CustomUserAccount.DoesNotExist:
+            raise CommandError('Current user does not exist')
+        
         try:
             with transaction.atomic():
                 user = CustomUserAccount.objects.create_user(username, first_name, last_name, email, phone_number, password, team_name)
@@ -75,13 +92,29 @@ def list_users(self):
 
 
     def create_superuser(self, options):
-        username = options['username'] or input('Enter username: ')
-        email = options['email'] or input('Enter email: ')
-        first_name = options['first_name'] or input('Enter first name: ')
-        last_name = options['last_name'] or input('Enter last name: ')
-        phone_number = options['phone_number'] or input('Enter phone number: ')
+        current_user_name = options['current_user'] or input(USER_DESCRIPTIONS['current_user'])
+        username = options['username'] or input(USER_DESCRIPTIONS['username'])
+        email = options['email'] or input(USER_DESCRIPTIONS['email'])
+        first_name = options['first_name'] or input(USER_DESCRIPTIONS['first_name'])
+        last_name = options['last_name'] or input(USER_DESCRIPTIONS['last_name'])
+        phone_number = options['phone_number'] or input(USER_DESCRIPTIONS['phone_number'])
         password = getpass('Enter password: ')
-        team_name = options['team_name'] or input('Enter team name: ')
+        team_name = (options['team_name'] or input(USER_DESCRIPTIONS['team_name'])).lower()
+        while team_name not in valid_team_names:
+            print(f'Invalid team name: {team_name}. Valid team_name options are : {", ".join(valid_team_names)}')
+            team_name = input(USER_DESCRIPTIONS['team_name']).lower()
+        
+        try:
+            current_user = CustomUserAccount.objects.get(username=current_user_name)
+            permission = IsAuthenticated(current_user).has_permission() and IsSuperuser(current_user).has_permission()
+            if not permission:
+                if not IsAuthenticated(current_user).has_permission():
+                    raise CommandError('Current user is not authenticated')
+                elif not IsSuperuser(current_user).has_permission() and CustomUserAccount.objects.filter(is_superuser=True).exists():
+                    raise CommandError('Current user is not a superuser and does not have permission to create a new user')
+        except CustomUserAccount.DoesNotExist:
+            raise CommandError('Current user does not exist')
+        
         try:
             with transaction.atomic():
                 user = CustomUserAccount.objects.create_superuser(username, email, first_name, last_name, phone_number, password, team_name)
@@ -91,36 +124,59 @@ def list_users(self):
 
 
     def delete_user(self, options):
+        current_user_name = options['current_user'] or input(USER_DESCRIPTIONS['current_user'])
         username = options['username'] or input('Enter username of user to delete: ')
         try:
+            current_user = CustomUserAccount.objects.get(username=current_user_name)
             user = CustomUserAccount.objects.get(username=username)
-            user.delete()
-            self.stdout.write(f'Successfully deleted user {username}')
+            permission = IsAuthenticated(current_user).has_permission() and (IsSuperuser(current_user).has_permission() or IsSameUser(user, current_user))
+            
+            if permission:
+                user.delete()
+                self.stdout.write(f'Successfully deleted user {username}')
         except CustomUserAccount.DoesNotExist:
             raise CommandError('User does not exist')
 
     def update_user(self, options):
+        current_user_name = options['current_user'] or input(USER_DESCRIPTIONS['current_user'])
         username = options['username'] or input('Enter username of user to update: ')
         try:
+            current_user = CustomUserAccount.objects.get(username=current_user_name)
             user = CustomUserAccount.objects.get(username=username)
-            for field in ['email', 'first_name', 'last_name', 'phone_number', 'team_name']:
-                value = options[field] or input(f'Enter new {field}: ')
-                if field == 'team_name':
-                    team = Team.objects.get(name=value)
-                    user.team = team
-                else:
-                    setattr(user, field, value)
-            user.save()
-            self.stdout.write(f'Successfully updated user {user}')
+            Permission = IsAuthenticated(current_user).has_permission() and (IsSuperuser(current_user).has_permission() or IsSameUser(user, current_user))
+            
+            if Permission:
+                with transaction.atomic():
+                    for field in ['email', 'first_name', 'last_name', 'phone_number', 'team_name']:
+                        value = options[field] or input(f'Enter new {field}: ')
+                        if valut: # Check if value is not an empty string
+                            if field == 'team_name':
+                                team_name = value.lower()
+                                while team_name not in valid_team_names:
+                                    print(f'Invalid team name: {team_name}. Valid team_name options are : {", ".join(valid_team_names)}')
+                                    team_name = input(USER_DESCRIPTIONS['team_name']).lower()
+                                try:
+                                    team = Team.objects.get(name=team_name)
+                                except Team.DoesNotExist:
+                                    raise CommandError(f'Team {team_name} does not exist')
+                                user.team = team
+                            else:
+                                setattr(user, field, value)
+                    user.save()
+                    self.stdout.write(f'Successfully updated user {user}')
         except CustomUserAccount.DoesNotExist:
             raise CommandError('User does not exist')
-        except Team.DoesNotExist:
-            raise CommandError('Team does not exist')
 
     def read_user(self, options):
+        current_user_name = options['current_user'] or input(USER_DESCRIPTIONS['current_user'])
         username = options['username'] or input('Enter username of user to read: ')
         try:
-            user = CustomUserAccount.objects.get(username=username)
-            self.stdout.write(str(user))
+            current_user = CustomUserAccount.objects.get(username=current_user_name)
+            permission = IsAuthenticated(current_user).has_permission()
+            
+            if permission:
+                user = CustomUserAccount.objects.get(username=username)
+                user_info = f"Username: {user.username}\n, Last Name: {user.last_name}\n, First Name: {user.first_name}\n, Email: {user.email}, Team Name: {user.team.name}\n, Date Joined: {user.date_joined}\n, Last Login: {user.last_login}\n"
+                self.stdout.write(user_info)
         except CustomUserAccount.DoesNotExist:
             raise CommandError('User does not exist')
